load works. testing count words

+++ big_data/load_bz2.py
@@ -22,7 +22,7 @@ def download(url, file_out="download.out"):
     # From http://stackoverflow.com/questions/4028697
     # /how-do-i-download-a-zip-file-in-python-using-urllib2
     try:
-        print("Downloading\n{url}\nto\n{file_out}".format(url=url, file_out=file_out))
+        print("Downloading:\n{url}\nto\n{file_out}".format(url=url, file_out=file_out))
         f_url = urllib2.urlopen(url)
         with open(file_out, 'wb') as f_out:
             f_out.write(f_url.read())
@@ -44,7 +44,7 @@ def decompress_and_partition(file_bz2, file_out="decompress.out"):
         print(file_bz2)
         raise NameError("File extension not '.bz2'.")
     print(
-        "Decompressing and partitioning\n"
+        "Decompressing and partitioning:\n"
         +"{file_bz2}\nto\n{file_out}".format(file_bz2=file_bz2, file_out=file_out))
     # From http://stackoverflow.com/questions/16963352/decompress-bz2-files
     # and http://bob.ippoli.to/archives/2005/06/14/python-iterators-and-sentinel-values/
@@ -54,19 +54,16 @@ def decompress_and_partition(file_bz2, file_out="decompress.out"):
             f_out.write(b"\n")
     return None
 
-def main(file_in="bz2_url_list.txt", tmp_dir="/scratch/sth499", test=False):
+def main(file_in, tmp_dir, tag, delete):
     """
-    Download bz2 files from list and upload to Disco Distributed File System.
+    Download bz2 files from list and upload to Disco.
     """
     with open(file_in, 'r') as f_in:
+
         # If Disco tag exists, delete it.
-        if test:
-            tag="data:big_test"
-        else:
-            tag="data:big"
-            # TODO: RESUME HERE WITH CHOOSING TEST FILE TO LOAD. make tag an arg. 
+        # Don't add all-new data to an already existing tag.
         if DDFS().exists(tag=tag):
-            print("Deleting Disco tag {tag}.".format(tag=tag))
+            print("WARNING: Overwriting Disco tag {tag}.".format(tag=tag), file=sys.stderr)
             DDFS().delete(tag=tag)
 
         for line in f_in:
@@ -78,23 +75,34 @@ def main(file_in="bz2_url_list.txt", tmp_dir="/scratch/sth499", test=False):
             # Remove newlines and name file from URL.
             url = line.strip()
             file_bz2 = os.path.join(tmp_dir, os.path.basename(url))
-            # Download bz2 file.
-            download(url=url, file_out=file_bz2)
+            # Download bz2 file if it doesn't exist.
+            if os.path.isfile(file_bz2):
+                print("Skipping download. File already exists:\n{file_bz2}".format(file_bz2=file_bz2))
+            else:
+                download(url=url, file_out=file_bz2)
 
-            # Decompress and partition bz2 file.
+            # Decompress and partition bz2 file if it doesn't exist.
             file_decom = os.path.splitext(file_bz2)[0]
-            decompress_and_partition(file_bz2=file_bz2, file_out=file_decom)
+            if os.path.isfile(file_decom):
+                print("Skipping decompress and partition. "
+                      +"File already exists:\n{file_decom}".format(file_decom=file_decom))
+            else:
+                decompress_and_partition(file_bz2=file_bz2, file_out=file_decom)
 
             # Load data into Disco Distributed File System.
             print("Loading into Disco:\n{file_in}".format(file_in=file_decom))
             try:
                 DDFS().chunk(tag=tag, urls=[os.path.join('./', file_decom)])
-            except ValueError:
-                print("ValueError:\n{file_decom}".format(file_decom=file_decom), file=sys.stderr)
+            except ValueError as err:
+                print("ValueError: " + err.message, file=sys.stderr)
+                print("File: {file_decom}".format(file_decom=file_decom), file=sys.stderr)
+
             # Delete bz2 and decompressed files.
-            print("Deleting:\n{file_bz2}\n{file_decom}".format(file_bz2=file_bz2, file_decom=file_decom))
-            os.remove(file_bz2)
-            os.remove(file_decom)
+            if delete:
+                print("Deleting:\n{file_bz2}\n{file_decom}".format(file_bz2=file_bz2, file_decom=file_decom))
+                os.remove(file_bz2)
+                os.remove(file_decom)
+
     return None
 
 if __name__ == '__main__':
@@ -102,15 +110,18 @@ if __name__ == '__main__':
     file_in_default = "bz2_url_list.txt"
     tmp_dir_default = "/tmp"
     tag_default = "data:big"
+    (file_in_base, ext) = os.path.splitext(file_in_default)
+    file_in_test = file_in_base + "_test" + ext
+    tag_test = tag_default + "_test"
     
-    parser = argparse.ArgumentParser(description="Download bz2 files from list, upload to Disco and tag.")
+    parser = argparse.ArgumentParser(description="Download bz2 files from list then upload to Disco and tag.")
     parser.add_argument("--file_in",
                         default=file_in_default, 
                         help=("Input file list of URLs to bz2 files for download. "
                               +"Default: {default}".format(default=file_in_default)))
     parser.add_argument("--tmp_dir",
                         default=tmp_dir_default,
-                        help=("Path to temporarily save bz2 files for extraction and loading. "
+                        help=("Path to save bz2 files for extraction and loading. "
                               +"Default: {default}".format(default=tmp_dir_default)))
     parser.add_argument("--tag",
                         default=tag_default,
@@ -118,12 +129,17 @@ if __name__ == '__main__':
                               +"Default: {default}".format(default=tag_default)))
     parser.add_argument("--test",
                         action="store_true",
-                        help=("T/F flag to load smaller 1GB data set from {file_test} to {tag_test}."
-                              .format(file_test=file_in_default+"_test", tag_test=tag_default+"_test")))
+                        help=("T/F flag to load smaller data set from {file} to {tag}."
+                              .format(file=file_in_test, tag=tag_test)))
+    parser.add_argument("--delete",
+                        action="store_true",
+                        help="T/F flag to delete files after uploaded to Disco.")
     args = parser.parse_args()
-    print(args)
 
     if args.test:
-        file_in = args.file
+        args.file_in = file_in_test
+        args.tag = tag_test
 
-    main(file_in=args.file_in, tmp_dir=args.tmp_dir, tag=args.tag, test=args.test)
+    print(args)
+
+    main(file_in=args.file_in, tmp_dir=args.tmp_dir, tag=args.tag, delete=args.delete)
